requirements.txt
anyio==3.6.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dateparser==1.1.1
fastapi==0.85.1
future==0.18.2
GoogleNews==1.6.5
h11==0.14.0
idna==3.4
lxml==4.9.1
numpy==1.23.4
oauthlib==3.2.2
pandas==1.5.0
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-twitter==3.5
pytrends==4.8.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.20.4
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.18.3
